# perf(remix-dev/vite): use `transformRequest` for route exports anlaysis to utilize Vite's caching