--- a/packages/remix-dev/vite/plugin.ts
+++ b/packages/remix-dev/vite/plugin.ts
@@ -239,9 +239,7 @@ const getRouteModuleExports = async (
   let routePath = path.join(pluginConfig.appDirectory, routeFile);
   let url = resolveFileUrl(pluginConfig, routePath);
 
-  let transformed = await viteChildCompiler.transformRequest(url, {
-    ssr: true,
-  });
+  let transformed = await viteChildCompiler.transformRequest(url);
   invariant(transformed, "Failed to transform: " + routeFile);
   let [, exports] = esModuleLexer(transformed.code);
   let exportNames = exports.map((e) => e.n);
@@ -649,9 +647,6 @@ export const remixVitePlugin: RemixVitePlugin = (options = {}) => {
 
         viteChildCompiler = await vite.createServer({
           ...viteUserConfig,
-          experimental: {
-            skipSsrTransform: true,
-          },
           mode: viteConfig.mode,
           server: {
             preTransformRequests: false,