--- a/packages/remix-dev/vite/plugin.ts
+++ b/packages/remix-dev/vite/plugin.ts
@@ -205,7 +205,21 @@ const writeFileSafe = async (file: string, contents: string): Promise<void> => {
   await fse.writeFile(file, contents);
 };
 
-const getRouteManifestModuleExports = async (
+function wrapConsoleTimeAsync<F extends (...args: any[]) => any>(f: F) {
+  const wrapper = async function (this: any, ...args: any[]) {
+    const id = Math.random().toString(36).slice(2).padStart(12, "0");
+    const label = `${f.name}:${id}`;
+    globalThis.console.time(label);
+    try {
+      return await f.apply(this, args);
+    } finally {
+      globalThis.console.timeEnd(label);
+    }
+  }
+  return wrapper as F;
+}
+
+let getRouteManifestModuleExports = async (
   viteChildCompiler: Vite.ViteDevServer | null,
   pluginConfig: ResolvedRemixVitePluginConfig
 ): Promise<Record<string, string[]>> => {
@@ -222,7 +236,9 @@ const getRouteManifestModuleExports = async (
   return Object.fromEntries(entries);
 };
 
-const getRouteModuleExports = async (
+getRouteManifestModuleExports = wrapConsoleTimeAsync(getRouteManifestModuleExports);
+
+let getRouteModuleExports = async (
   viteChildCompiler: Vite.ViteDevServer | null,
   pluginConfig: ResolvedRemixVitePluginConfig,
   routeFile: string
@@ -246,6 +262,8 @@ const getRouteModuleExports = async (
 
   return exportNames;
 };
+
+getRouteModuleExports = wrapConsoleTimeAsync(getRouteModuleExports);
 
 const showUnstableWarning = () => {
   console.warn(