@hi-ogawa 2023-11-28
I thought this idea makes sense and it feels like a nice cleanup as well, but I don't have a definitive way to measure overall perf impact, so let me close this for now.

I'm suspecting the actual bottleneck is `ssrLoadModule` of server build, so saving some time on repeated calls of child compiler's `getRouteManifestModuleExports` might not be so effective.